// npm/yuku-ast/src/generated.ts
const TYPES = [
  "SequenceExpression",
  "ParenthesizedExpression",
  "ArrowFunctionExpression",
  "FunctionDeclaration",
  "FunctionExpression",
  "TSDeclareFunction",
  "TSEmptyBodyFunctionExpression",
  "BlockStatement",
  "BlockStatement",
  "BinaryExpression",
  "LogicalExpression",
  "ConditionalExpression",
  "UnaryExpression",
  "UpdateExpression",
  "AssignmentExpression",
  "ArrayExpression",
  "ObjectExpression",
  "SpreadElement",
  "Property",
  "MemberExpression",
  "CallExpression",
  "ChainExpression",
  "TaggedTemplateExpression",
  "NewExpression",
  "AwaitExpression",
  "YieldExpression",
  "MetaProperty",
  "Decorator",
  "ClassDeclaration",
  "ClassExpression",
  "ClassBody",
  "MethodDefinition",
  "TSAbstractMethodDefinition",
  "PropertyDefinition",
  "AccessorProperty",
  "TSAbstractPropertyDefinition",
  "TSAbstractAccessorProperty",
  "StaticBlock",
  "Super",
  "Literal",
  "Literal",
  "Literal",
  "Literal",
  "Literal",
  "ThisExpression",
  "Literal",
  "TemplateLiteral",
  "TemplateElement",
  "Identifier",
  "PrivateIdentifier",
  "Identifier",
  "Identifier",
  "Identifier",
  "ExpressionStatement",
  "IfStatement",
  "SwitchStatement",
  "SwitchCase",
  "ForStatement",
  "ForInStatement",
  "ForOfStatement",
  "WhileStatement",
  "DoWhileStatement",
  "BreakStatement",
  "ContinueStatement",
  "LabeledStatement",
  "WithStatement",
  "ReturnStatement",
  "ThrowStatement",
  "TryStatement",
  "CatchClause",
  "DebuggerStatement",
  "EmptyStatement",
  "VariableDeclaration",
  "VariableDeclarator",
  "ExpressionStatement",
  "AssignmentPattern",
  "RestElement",
  "ArrayPattern",
  "ObjectPattern",
  "Property",
  "Program",
  "ImportExpression",
  "ImportDeclaration",
  "ImportSpecifier",
  "ImportDefaultSpecifier",
  "ImportNamespaceSpecifier",
  "ImportAttribute",
  "ExportNamedDeclaration",
  "ExportDefaultDeclaration",
  "ExportAllDeclaration",
  "ExportSpecifier",
  "TSTypeAnnotation",
  "TSAnyKeyword",
  "TSUnknownKeyword",
  "TSNeverKeyword",
  "TSVoidKeyword",
  "TSNullKeyword",
  "TSUndefinedKeyword",
  "TSStringKeyword",
  "TSNumberKeyword",
  "TSBigIntKeyword",
  "TSBooleanKeyword",
  "TSSymbolKeyword",
  "TSObjectKeyword",
  "TSIntrinsicKeyword",
  "TSThisType",
  "TSTypeReference",
  "TSQualifiedName",
  "TSTypeQuery",
  "TSImportType",
  "TSTypeParameter",
  "TSTypeParameterDeclaration",
  "TSTypeParameterInstantiation",
  "TSLiteralType",
  "TSTemplateLiteralType",
  "TSArrayType",
  "TSIndexedAccessType",
  "TSTupleType",
  "TSNamedTupleMember",
  "TSOptionalType",
  "TSRestType",
  "TSJSDocNullableType",
  "TSJSDocNonNullableType",
  "TSJSDocUnknownType",
  "TSUnionType",
  "TSIntersectionType",
  "TSConditionalType",
  "TSInferType",
  "TSTypeOperator",
  "TSParenthesizedType",
  "TSFunctionType",
  "TSConstructorType",
  "TSTypePredicate",
  "TSTypeLiteral",
  "TSMappedType",
  "TSPropertySignature",
  "TSMethodSignature",
  "TSCallSignatureDeclaration",
  "TSConstructSignatureDeclaration",
  "TSIndexSignature",
  "TSTypeAliasDeclaration",
  "TSInterfaceDeclaration",
  "TSInterfaceBody",
  "TSInterfaceHeritage",
  "TSClassImplements",
  "TSEnumDeclaration",
  "TSEnumBody",
  "TSEnumMember",
  "TSModuleDeclaration",
  "TSModuleBlock",
  "TSModuleDeclaration",
  "TSParameterProperty",
  "Identifier",
  "TSAsExpression",
  "TSSatisfiesExpression",
  "TSTypeAssertion",
  "TSNonNullExpression",
  "TSInstantiationExpression",
  "TSExportAssignment",
  "TSNamespaceExportDeclaration",
  "TSImportEqualsDeclaration",
  "TSExternalModuleReference",
  "JSXElement",
  "JSXOpeningElement",
  "JSXClosingElement",
  "JSXFragment",
  "JSXOpeningFragment",
  "JSXClosingFragment",
  "JSXIdentifier",
  "JSXNamespacedName",
  "JSXMemberExpression",
  "JSXAttribute",
  "JSXSpreadAttribute",
  "JSXExpressionContainer",
  "JSXEmptyExpression",
  "JSXText",
  "JSXSpreadChild",
  "Hashbang",
] as const;

export type GeneratedNodeType = (typeof TYPES)[number];

export const NODE_TYPES: readonly GeneratedNodeType[] = [...new Set<GeneratedNodeType>(TYPES)];